+++ packages/gatsby-theme-availity/gatsby-config.js
@@ -1,3 +1,22 @@
+const defaultNavConfig = {
+  '/availity-workflow': {
+    text: 'Getting Started',
+    matchRegex: '^/availity-workflow',
+  },
+  '/availity-uikit': {
+    text: 'UI Kit',
+    matchRegex: '^/availity-uikit',
+  },
+  '/availity-react': {
+    text: 'Components',
+    matchRegex: '^/availity-react',
+  },
+  '/sdk-js': {
+    text: 'Resources',
+    matchRegex: '^/sdk-js',
+  },
+};
+
 module.exports = ({
   root,
   subtitle,