+++ packages/gatsby-theme-availity/src/index.js
@@ -28,16 +28,14 @@ const renderAst = new RehypeReact({
 
 // The Template to load on each page
 const Template = ({
-  children,
   location,
-  pageContext: { sidebarContents },
+  pageContext: { sidebarContents, navItems },
   data,
-  ...rest
 }) => {
   // Keep a ref of the current content window for jumping to certain anchors in section nav
   const mainRef = useRef(null);
 
-  const { file } = data;
+  const { file, site } = data;
   const { frontmatter, headings } = file.childMarkdownRemark || file.childMdx;
 
   const { hash, pathname } = location;
@@ -46,15 +44,25 @@ const Template = ({
     .reduce((acc, { pages }) => acc.concat(pages), [])
     .filter(page => !page.anchor);
 
+  function isPathActive(value) {
+    return !location.pathname.indexOf(value);
+  }
+
   return (
     <div className="h-100 d-flex flex-column">
       <SiteMetadata pathname={pathname} />
-      <Navigation />
+      <Navigation
+        className="pl-4"
+        navItems={navItems}
+        pathname={pathname}
+        isPathActive={isPathActive}
+      />
       <div className="d-flex h-100">
         <SideNavigation
-          currentPath={rest.path}
+          currentPath={pathname}
           contents={sidebarContents}
-          className="flex-shrink-0 ml-5 pt-5"
+          siteTitle={site.siteMetadata.subtitle}
+          className="flex-shrink-0 ml-4 pt-4"
           style={{
             overflowY: 'auto',
             width: 200,
@@ -97,6 +105,7 @@ export const pageQuery = graphql`
     site {
       siteMetadata {
         title
+        subtitle
         description
       }
     }