+++ packages/gatsby-theme-availity/theme-options.js
@@ -1,0 +1,24 @@
+const navConfig = {
+  '/availity-workflow': {
+    text: 'Getting Started',
+    matchRegex: '^/availity-workflow',
+  },
+  '/availity-uikit': {
+    text: 'UI Kit',
+    matchRegex: '^/availity-uikit',
+  },
+  '/availity-react': {
+    text: 'Components',
+    matchRegex: '^/availity-react',
+  },
+  '/sdk-js': {
+    text: 'Resources',
+    matchRegex: '^/sdk-js',
+  },
+};
+
+module.exports = {
+  siteName: 'Availity Docs',
+  baseUrl: 'https://availity.github.io',
+  navConfig,
+};