+++ packages/site/gatsby-config.js
@@ -1,16 +1,19 @@
+const themeOptions = require('@availity/gatsby-theme/theme-options');
+
 module.exports = {
   pathPrefix: '/docs/sdk',
   __experimentalThemes: [
     {
       resolve: 'gatsby-theme-availity',
       options: {
+        ...themeOptions,
         root: __dirname,
         subtitle: 'Availity SDK',
         description:
           'Documentation for Availity Javascript SDK, Rest API and GraphQL',
         githubRepo: 'availity/sdk-js',
         sidebarCategories: {
-          'Getting Started': ['index'],
+          null: ['index'],
           'API Resources': [
             'api-axios/api-axios',
             'api-core/api-core',